RoboKidz/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.9.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-extensions==0.7.1
drf-yasg==1.21.3
ffmpy==0.3.0
idna==3.3
imageio==2.22.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
networkx==2.8.6
numpy==1.23.3
openapi-codec==1.3.2
opencv-python==4.6.0.66
packaging==21.3
pafy==0.5.5
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyotp==2.3.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scipy==1.9.1
simplejson==3.17.6
sqlparse==0.4.2
tifffile==2022.8.12
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
youtube-dl==2020.12.2
